chore: Refactor DECLARE statement
